requirements.txt
streamlit
langchain
envyaml
openai
python-dotenv
langchain_community
langchain-openai